Configured tile set for player 0 and did some minor restructuring
