+++ locstats.md
@@ -1,10 +1,10 @@
 
-cloc|github.com/AlDanial/cloc v 1.82  T=0.12 s (1016.3 files/s, 87759.0 lines/s)
+cloc|github.com/AlDanial/cloc v 1.82  T=0.07 s (1752.7 files/s, 151671.0 lines/s)
 --- | ---
 
 Language|files|blank|comment|code
 :-------|-------:|-------:|-------:|-------:
-TypeScript|84|810|200|8653
+TypeScript|85|814|201|8757
 CSS|27|78|0|528
 Markdown|7|55|0|150
 JavaScript|1|12|31|98
@@ -13,4 +13,4 @@ YAML|2|2|0|26
 JSON|1|0|0|22
 Bourne Shell|1|3|3|9
 --------|--------|--------|--------|--------
-SUM:|124|960|234|9514
+SUM:|125|964|235|9618
